requirements.txt
bleach==6.0.0
blinker==1.5
click==8.1.3
Deprecated==1.2.13
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
flask-csp==0.10
Flask-Limiter==1.4
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-SQLAlchemy==3.0.3
flask-talisman==1.0.0
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.3.1
MarkupSafe==2.1.2
packaging==23.0
passlib==1.7.4
psycopg2-binary==2.9.5
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
zipp==3.14.0